Moving into microservices folder, performing API error handling, config restructure, adding readme
