+++ microservices/google-api/google_search.json
@@ -1,8 +1,9 @@
 {
     "bucket": "sp-ca-bc-gov-131565110619-12-microservices",
-    "sites": [
-        "https://www2.gov.bc.ca/"
-    ],
     "dbtable": "google.googlesearch",
-    "start_date_default": "2017-11-09"
+    "sites":[
+        { 
+        "name":"https://www2.gov.bc.ca/"
+        }
+    ]
 }
